Return an empty hash for unknown hosts
When using a static inventory that declares new hosts alongside a
dynamic inventory provided by this utility, ansible will call the
dynamic inventory with `--host <host>` to merge in additional vars
that may have been provided by the dynamic inventory.

The default behaviour of *terraform-inventory* is to output an error
message when an unknown host is encountered, breaking this
functionality.

The correct behaviour is to return an empty hash implying there are
no additional vars for this host.