Add support for azurerm_virtual_machine
An azurerm_network_interface must be defined and associated with
the VM. Only supports loading private IP addresses. If the nic
has multiple IP addresses defined, it will use the first one
(read from private_ip_address). This is the primary IP address
according to documentation.

If there are multiple nics defined for the VM, the nic declared
as the primary nic (via primary_network_interface_id) will be used.