Terraform 0.12+ index can be string or number
In the case of when mixing resources into the statefile in terraform, it's now
possible to have indexes that are not always int. In the example below using the
new 0.12 syntax `foreach` will result in the index being mapped to the
`each.key` name instead of the integer indexes.

```
variable "test" {
  default = {"testa" = "1.1.1.1", "testb" = "2.2.2.2"}
}

resource "aws_route53_record" "this" {
  for_each = var.test
  zone_id = "abc"
  name    = "${each.key}.tv."
  type    = "A"
  ttl     = "300"
  records = [each.value]
}
```